fix cred data
